Clients that use "dnssec_chain" MUST send SNI.
Servers that don't support the client's SNI MUST decline the
extension.  In order to deliver (or show the non-existence of) some
alternate TLSA RRset, the server would have to disprove the existence
of the one the client requested.  And with denial of existence for
the alternative name the client may have to be able to figure which
name the server is denying, despite its absence in the returned DNS
records.  All this is much too complex and unnecessary, if the
client's SNI is not supported, the server can't employ this extension
for that name.