Lift RRset ordering restrictions
For non-wildcard positive replies, that contain the TLSA RRset for
the requested SNI base domain and port number (along with the shared
transport protocol), the placement of that RRset is not significant,
the client can easily identify it as the requested record wherever
it may appear in the sequence of RRsets.

For denial of existence replies, identifying the matching NSEC/NSEC3
records is potentially more difficult, but it should be noted that
even if placing the relevant NSEC/NSEC3 RRset "first" (that is after
any initial CNAMEs that lead to where the non-existent TLSA records
"would have been") does simplify processing, the same problem of
identifying the pertinent NSEC records also comes up in the case
of wildcard answer synthesis, because validation of the wildcard
requires proof of non-existence of the exact requested name.

The draft does not specify the placement of denial of existence
RRsets in support of wildcard synthesis.  Such synthesis can
happen more than once:

        *._tcp.example.com. IN CNAME _443._tcp.example.net.
        *._tcp.example.net. IN TLSA 3 1 1 ...

So the client software in any case needs to be able to search the
RRset "heap" to find the appropriate records.  For this the client's
software needs to arrange the records in a tree, and find the most
specific enclosing delegation (DS RRset).  With the zone cut in
hand, to find the appropriate NSEC3 records it can then examine
just those NSEC3 records which are directly below that zone's apex.
The NSEC case is similar.

Since simple linear bottom-up validation is not feasible, and a
tree generally needs to be constructed to perform the validation,
and ordering constraints that work in some generality are difficult
to describe, I believe that it is best to abandon the ordering
requirements.  Implementations that don't assume any particular
order, are less likely to make invalid assumptions when validating
the response.

One might even, in order to "grease the protocol", suggest that the
server should consider returning the records in random order!